+++ backend/math/src/functions/exponential.rs
@@ -1,5 +1,6 @@
 use crate::Number;
 
+#[derive(Debug, Copy, Clone)]
 pub struct ExponentialFunction {
     pub c: Number,
     pub a: Number,